Navbar moved to higher level
