+++ package.json
@@ -16,6 +16,7 @@
     "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
     "react-with-firebase-auth": "^1.3.11",
+    "recharts": "^2.1.8",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
