+++ src/components/Console/Console.css
@@ -3,6 +3,7 @@
   width: 100vw;
   display: flex;
   flex-direction: column;
+  margin-top: 5%;
 }
 .dashboard_nav{
   width: 100%;
