+++ src/components/Home/Home.css
@@ -7,66 +7,7 @@
   background-color: #2b6777;
 }
 
-.home-nav {
-  width: 100%;
-  height: 10%;
-  background-color: rgba(255, 255, 255, 1);
-  display: flex;
-  position: fixed;
-  justify-content: space-between;
-  align-items: center;
-  padding-left: 15px;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-  box-shadow: 0px 5px 15px rgba(0, 0, 0, 0.4);
-}
 
-.home-nav-left {
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  width: 40%;
-  height: 90%;
-}
-
-.home-nav-right {
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  width: 15%;
-  height: 90%;
-  margin-right: 2%;
-}
-
-.nav-button-normal {
-  width: 100%;
-  height: 100%;
-  border: none;
-  background-color: rgba(255, 255, 255, 0);
-  font-family: "Roboto", sans-serif !important;
-  font-size: 1rem;
-  margin-left: -40px;
-}
-
-.nav-button {
-  width: 100%;
-  height: 100%;
-  border: none;
-  background-color: rgba(255, 255, 255, 0);
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  font-family: "Roboto", sans-serif !important;
-  font-size: 1rem;
-  transition: border-bottom 1s ease;
-  color: black;
-  text-decoration: unset !important;
-}
-
-.nav-button:hover {
-  border-bottom: 1px solid black;
-}
 
 .brand-name {
   width: 100%;
@@ -78,13 +19,7 @@
   align-items: center;
 }
 
-.nav-logo {
-  width: 70%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
+
 
 .home-body {
   width: 90%;
